Refactor: Extract Navbar logic into a custom hook and update Firebase configuration to use environment variables
