+++ src/config/firebase.ts
@@ -3,12 +3,12 @@ import { getAuth } from "firebase/auth";
 import { getFirestore } from "firebase/firestore";
 
 const firebaseConfig = {
-  apiKey: "AIzaSyBgav6ZfJ0xI1wSeRKdntXCI2GS64ouNOg",
-  authDomain: "tourly-6e2a8.firebaseapp.com",
-  projectId: "tourly-6e2a8",
-  storageBucket: "tourly-6e2a8.appspot.com",
-  messagingSenderId: "217329461679",
-  appId: "1:217329461679:web:c7a8e1b2d3f4e5f6g7h8i9" // Este debe ser el appId real de tu proyecto
+  apiKey: import.meta.env.VITE_FIREBASE_API_KEY,
+  authDomain: import.meta.env.VITE_FIREBASE_AUTH_DOMAIN,
+  projectId: import.meta.env.VITE_FIREBASE_PROJECT_ID,
+  storageBucket: import.meta.env.VITE_FIREBASE_STORAGE_BUCKET,
+  messagingSenderId: import.meta.env.VITE_FIREBASE_MESSAGING_SENDER_ID,
+  appId: import.meta.env.VITE_FIREBASE_APP_ID
 };
 
 // Inicializar Firebase
